requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
geographiclib==2.0
geopy==2.4.1
gunicorn==21.2.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
urllib3==2.3.0
